@@ -3,9 +3,9 @@
   "version": "1.0.0",
   "description": "A simple http download proxy.",
   "engines": {
-    "node": "14.x"
+    "node": "18.x"
   },
-  "main": "app.js",
+  "main": "src/app.mjs",
   "scripts": {
     "start": "node .",
     "dev": "nodemon ."
@@ -16,7 +16,7 @@
     "@koa/router": "^8.0.8",
     "koa": "^2.11.0",
     "koa-static-cache": "^5.1.3",
-    "node-fetch": "^3.1.1"
+    "node-fetch": "^3.2.4"
   },
   "devDependencies": {
     "nodemon": "^2.0.4"
