@@ -1,15 +1,19 @@
-const Koa = require("koa");
-const static = require("koa-static-cache");
-const path = require("path");
-const router = require("./controllers");
+import Koa from "koa"
+import koaStaticCache from "koa-static-cache";
+import path from "node:path"
+import router from "./controller.mjs"
+import { getDirname } from "./utils.mjs";
 
 const PORT = Number(process.env.PORT) || 3000;
 const app = new Koa();
-const staticOpt = static({
-  dir: path.resolve("./static"),
+const staticOpt = koaStaticCache({
+  dir: path.resolve(getDirname(), "../static"),
   buffer: true,
   gzip: true,
   usePrecompiledGzip: true,
+  alias: {
+    '/': '/index.html'
+  }
 });
 
 async function errorHandler(ctx, next) {
